fix: Persist result should respect the machine id
