+++ control-plane/src/modules/observability/events.ts
@@ -9,6 +9,7 @@ export type EventTypes =
   | "jobReceived"
   | "jobStatusRequest"
   | "jobResulted"
+  | "jobResultedButNotPersisted"
   | "jobStalled"
   | "jobStalledTooManyTimes"
   | "jobRecovered"
