+++ control-plane/src/utilities/errors.ts
@@ -1,4 +1,6 @@
 export class NotFoundError extends Error {
+  statusCode: number = 404;
+
   constructor(message: string) {
     super(message);
     this.name = "NotFoundError";
